+++ pytests/test_testing.py
@@ -18,59 +18,59 @@ def test_ffwd_iter():
         next(it)
 
 
-def test_testing_source(now):
+def test_testing_source():
     inp = TestingSource(range(3))
-    part = inp.build_part(now, "iterable", None)
-    assert part.next_batch(now) == [0]
-    assert part.next_batch(now) == [1]
-    assert part.next_batch(now) == [2]
+    part = inp.build_part("iterable", None)
+    assert part.next_batch() == [0]
+    assert part.next_batch() == [1]
+    assert part.next_batch() == [2]
     with raises(StopIteration):
-        part.next_batch(now)
+        part.next_batch()
     part.close()
 
 
-def test_testing_source_resume_state(now):
+def test_testing_source_resume_state():
     inp = TestingSource(range(3))
-    part = inp.build_part(now, "iterable", None)
-    assert part.next_batch(now) == [0]
-    assert part.next_batch(now) == [1]
+    part = inp.build_part("iterable", None)
+    assert part.next_batch() == [0]
+    assert part.next_batch() == [1]
     resume_state = part.snapshot()
     assert resume_state == 2
-    assert part.next_batch(now) == [2]
+    assert part.next_batch() == [2]
     part.close()
 
     inp = TestingSource(range(3))
-    part = inp.build_part(now, "iterable", resume_state)
+    part = inp.build_part("iterable", resume_state)
     assert part.snapshot() == resume_state
-    assert part.next_batch(now) == [2]
+    assert part.next_batch() == [2]
     with raises(StopIteration):
-        part.next_batch(now)
+        part.next_batch()
     part.close()
 
 
-def test_testing_source_batch_size(now):
+def test_testing_source_batch_size():
     inp = TestingSource(range(5), batch_size=2)
-    part = inp.build_part(now, "iterable", None)
-    assert part.next_batch(now) == [0, 1]
-    assert part.next_batch(now) == [2, 3]
-    assert part.next_batch(now) == [4]
+    part = inp.build_part("iterable", None)
+    assert part.next_batch() == [0, 1]
+    assert part.next_batch() == [2, 3]
+    assert part.next_batch() == [4]
     part.close()
 
 
-def test_testing_source_eof(now):
+def test_testing_source_eof():
     inp = TestingSource([0, 1, 2, TestingSource.EOF(), 3, 4], batch_size=2)
-    part = inp.build_part(now, "iterable", None)
-    assert part.next_batch(now) == [0, 1]
-    assert part.next_batch(now) == [2]
+    part = inp.build_part("iterable", None)
+    assert part.next_batch() == [0, 1]
+    assert part.next_batch() == [2]
     with raises(StopIteration):
-        part.next_batch(now)
+        part.next_batch()
     part.close()
 
     resume_state = part.snapshot()
-    part = inp.build_part(now, "iterable", resume_state)
-    assert part.next_batch(now) == [3, 4]
+    part = inp.build_part("iterable", resume_state)
+    assert part.next_batch() == [3, 4]
     with raises(StopIteration):
-        part.next_batch(now)
+        part.next_batch()
     part.close()
 
 
@@ -90,20 +90,20 @@ def test_testing_source_eof_run(recovery_config):
     assert out == [3, 4]
 
 
-def test_testing_source_abort(now):
+def test_testing_source_abort():
     inp = TestingSource([0, 1, 2, TestingSource.ABORT(), 3, 4], batch_size=2)
-    part = inp.build_part(now, "iterable", None)
-    assert part.next_batch(now) == [0, 1]
+    part = inp.build_part("iterable", None)
+    assert part.next_batch() == [0, 1]
     resume_state = part.snapshot()
-    assert part.next_batch(now) == [2]
+    assert part.next_batch() == [2]
     with raises(AbortExecution):
-        part.next_batch(now)
+        part.next_batch()
 
-    part = inp.build_part(now, "iterable", resume_state)
-    assert part.next_batch(now) == [2, 3]
-    assert part.next_batch(now) == [4]
+    part = inp.build_part("iterable", resume_state)
+    assert part.next_batch() == [2, 3]
+    assert part.next_batch() == [4]
     with raises(StopIteration):
-        part.next_batch(now)
+        part.next_batch()
     part.close()